+++ Listener/LogoutListener.php
@@ -9,29 +9,41 @@
 
 namespace c975L\UserBundle\Listener;
 
+use Doctrine\ORM\EntityManagerInterface;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\Security\Core\Authentication\Token\TokenInterface;
 use Symfony\Component\Security\Http\Logout\LogoutHandlerInterface;
 use c975L\UserBundle\Entity\UserAbstract;
 
+/**
+ * Class to listen to logout event
+ * @author Laurent Marquet <laurent.marquet@laposte.net>
+ * @copyright 2018 975L <contact@975l.com>
+ */
 class LogoutListener implements LogoutHandlerInterface
 {
+    /**
+     * Stores EntityManagerInterface
+     * @var EntityManagerInterface
+     */
     private $em;
 
-    public function __construct(\Doctrine\ORM\EntityManagerInterface $em)
+    public function __construct(EntityManagerInterface $em)
     {
         $this->em = $em;
     }
 
+    /**
+     * Adds data to user entity and persists
+     */
     public function logout(Request $request, Response $response, TokenInterface $token)
     {
-        //Gets the user
         $user = $token->getUser();
-
         if (is_subclass_of($user, 'c975L\UserBundle\Entity\UserAbstract')) {
             //Writes signout time
             $user->setLatestSignout(new \DateTime());
+
             $this->em->persist($user);
             $this->em->flush();
         }